Update mesa-demos to 8.0.1.
This is the latest release of mesa-demos, allthough it is already ancient.  It
gives us a place to start if we want to roll our own snapshots however.  Some
testing might be needed, especially when  using WITH_NVIDIA, since I lack such
hardware.  Some wordsmithing might also be good in pkg-descr.



git-svn-id: https://trillian.chruetertee.ch/svn/ports/trunk@421 058c260c-8361-11dd-a0ac-aa2bafec7d09