Merge r822-823:
Try to fix build of imake and ports using imake on 10-CURRENT where clang is
installed as cc.  This is done by hard-coding the use of gcpp in those
instances.
This means that if gcpp isn't installed this will fail, and it is not
possible to change the cpp imake uses without hacking bsd.port.mk or
similar.

Submitted by:	dim, kwm, zeising


git-svn-id: https://trillian.chruetertee.ch/svn/ports/branches/xorg-7.7@824 058c260c-8361-11dd-a0ac-aa2bafec7d09