graphics/libGL: Use "rm -f" to delete dri.pc (same as r1443 in trunk)
N-i-c-o on #freebsd-fr reported that dri.pc doesn't exist in the stage
directory on ARM. Thus, calling rm(1) on this non-existing file fails.
Passing "-f" to rm(1) fixes the problem.

Because the same commands are executed in both branches of the "if
(WITH_NEW_XORG)" block, move them outside of it.


git-svn-id: https://trillian.chruetertee.ch/svn/ports/branches/experimental@1446 058c260c-8361-11dd-a0ac-aa2bafec7d09