x11-drivers/xf86-input-keyboard: Don't query console keymap
This seems to be useless and doesn't work anymore with vt(4), due to the
use of Unicode codepoints in vt(4). The new patch remove the related
code. This fixes a segmentation fault during X.Org server startup.

This is a merge from trunk of r1506.

PR:		191459
Submitted by:	Claude Buisson <clbuisson@orange.fr> (earlier version)


git-svn-id: https://trillian.chruetertee.ch/svn/ports/branches/experimental@1509 058c260c-8361-11dd-a0ac-aa2bafec7d09