Import Mesa work from experimental to trunk.
Features:
Default mesa updated to 9.1.7.
New mesa updated to 10.3.0.
Drop usage of WITH_NEW_XORG, instead have WITH_NEW_MESA. This is not
really intended as a user operated flag. It is used to switch to mesa
10.3.0 when i915kms hw context support is found based on OSVERSION.
Radeon user might set this because radeonkms works fine with the new
mesa.


git-svn-id: https://trillian.chruetertee.ch/svn/ports/trunk@1529 058c260c-8361-11dd-a0ac-aa2bafec7d09